refactor(testing): moved mock client out of source
